requirements.txt
PyDispatcher
torch
TTS
EbookLib
bs4
tinydb
soundfile
numpy